backend/pyproject.toml
[project]
name = "github-chat"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "embedchain>=0.1.127",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "gitpython==3.1.38",
    "langchain-groq>=0.2.5",
    "langchain-huggingface>=0.1.2",
    "pygithub==1.59.1",
    "python-dotenv>=1.0.1",
    "sentence-transformers>=5.1.1",
]
